feat(fixed): goToAddress when there is an address
